evaluation/requirements.txt
datasets==4.5.0
tqdm
pandas==1.3.5
PyYAML
Pillow==9.5.0
openai==2.28.0
huggingface-hub==1.7.1
